robot-map-gui files renamed according to codestyle
