+++ apps/robot-map-gui/gui/CMainWindow.cpp
