+++ apps/robot-map-gui/gui/CMainWindow.h
@@ -2,7 +2,6 @@
 #include <QMainWindow>
 
 
-
 class CMainWindow : public QMainWindow {
 	Q_OBJECT
