+++ cmakemodules/script_show_final_summary.cmake
@@ -121,20 +121,21 @@ SHOW_CONFIG_LINE_SYSTEM("Assimp (3D models)                  " CMAKE_MRPT_HAS_AS
 SHOW_CONFIG_LINE_SYSTEM("eigen3                              " CMAKE_MRPT_HAS_EIGEN "[Version: ${MRPT_EIGEN_VERSION}]")
 SHOW_CONFIG_LINE_SYSTEM("ffmpeg libs (Video streaming)       " CMAKE_MRPT_HAS_FFMPEG "[avcodec ${LIBAVCODEC_VERSION}, avutil ${LIBAVUTIL_VERSION}, avformat ${LIBAVFORMAT_VERSION}]")
 SHOW_CONFIG_LINE_SYSTEM("gtest (Google unit testing library) " CMAKE_MRPT_HAS_GTEST )
-SHOW_CONFIG_LINE("Intel threading lib (TBB)           " CMAKE_MRPT_HAS_TBB)
+SHOW_CONFIG_LINE("Intel threading lib (TBB)                  " CMAKE_MRPT_HAS_TBB)
 SHOW_CONFIG_LINE_SYSTEM("lib3ds (3DStudio scenes)            " CMAKE_MRPT_HAS_LIB3DS)
 SHOW_CONFIG_LINE_SYSTEM("libclang (for ConvertUTF)           " CMAKE_MRPT_HAS_CLANG)
 SHOW_CONFIG_LINE_SYSTEM("libjpeg (jpeg)                      " CMAKE_MRPT_HAS_JPEG)
 SHOW_CONFIG_LINE_SYSTEM("liblas (ASPRS LAS LiDAR format)     " CMAKE_MRPT_HAS_LIBLAS)
-SHOW_CONFIG_LINE("mexplus                             " CMAKE_MRPT_HAS_MATLAB)
+SHOW_CONFIG_LINE("mexplus                                    " CMAKE_MRPT_HAS_MATLAB)
 SHOW_CONFIG_LINE_SYSTEM("OpenCV (Image manipulation)         " CMAKE_MRPT_HAS_OPENCV "[Version: ${MRPT_OPENCV_VERSION}]")
 SHOW_CONFIG_LINE_SYSTEM("OpenGL                              " CMAKE_MRPT_HAS_OPENGL_GLUT)
 SHOW_CONFIG_LINE_SYSTEM("GLUT                                " CMAKE_MRPT_HAS_GLUT)
 SHOW_CONFIG_LINE_SYSTEM("PCAP (Wireshark logs for Velodyne)  " CMAKE_MRPT_HAS_LIBPCAP)
 SHOW_CONFIG_LINE_SYSTEM("PCL (Pointscloud library)           " CMAKE_MRPT_HAS_PCL  "[Version: ${PCL_VERSION}]")
-SHOW_CONFIG_LINE("SuiteSparse                         " CMAKE_MRPT_HAS_SUITESPARSE)
-SHOW_CONFIG_LINE("VTK                                 " CMAKE_MRPT_HAS_VTK)
+SHOW_CONFIG_LINE("SuiteSparse                                " CMAKE_MRPT_HAS_SUITESPARSE)
+SHOW_CONFIG_LINE("VTK                                        " CMAKE_MRPT_HAS_VTK)
 SHOW_CONFIG_LINE_SYSTEM("wxWidgets                           " CMAKE_MRPT_HAS_WXWIDGETS)
+SHOW_CONFIG_LINE_SYSTEM("Qt5                                 " CMAKE_MRPT_HAS_Qt5)
 SHOW_CONFIG_LINE_SYSTEM("zlib (compression)                  " CMAKE_MRPT_HAS_ZLIB)
 SHOW_CONFIG_LINE_SYSTEM("yamlcpp (YAML file format)          " CMAKE_MRPT_HAS_YAMLCPP "[Version: ${LIBYAMLCPP_VERSION}]")
 MESSAGE(STATUS  "")
